Update __init__.py
Removed the reliance on the zip file and its unnecessary extraction every single time the user unwraps an object. Now the user sets the path to the directory where MinistryOfFlat is extracted instead. Also made it so the current object is still selected after unwrapping, this way the user doesn't have to switch back and forth between Object and Edit mode to re-unwrap and can see the changes on the fly.